+++ NEWS.rst
@@ -5,6 +5,8 @@ NEWS for system-image updater
 1.3 (2013-XX-XX)
 ================
  * Fixed bug in resolving channels with dashes in their name. (LP: #1217932)
+ * Add `system-image-cli --filter` option to allow for forcing full or delta
+   updates.  (LP: #1208909)
 
 1.2 (2013-08-26)
 ================